# docs/openapi/charge-console-v1.5.3.yaml
openapi: 3.0.3
info:
  title: Charge Console API
  version: 1.5.3
  description: |
    Charge Console 1.5.3 添加前绑定引导、扫码确认与现有整桩提醒契约。
    POST /api/piles 沿用现有字段；YYB_BINDING_REQUIRED 和 YYB_RESCAN_REQUIRED 返回 409，可就地引导绑定。
    临时提醒仅在用户指定时长内按桩号低频检查，首次发现空闲并生成通知后自动结束。
    WxPusher 是附加发送方式；通知中心始终保存消息，平台成功发送不代表特定客户端已读。
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
...
  /api/session/yyb-qr:
    post:
      operationId: createYybQr
      summary: 创建归属当前用户的短期扫码会话
      responses:
        "200":
          description: 后端重启或 30 分钟后需重新生成；不保存绑定
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YybQr"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "503":
          description: 未配置扫码或暂时无法创建会话
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: charge_session
...
  responses:
...
    Unauthenticated:
      description: 未登录或会话已失效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: login required
...
  schemas:
...
    YybQr:
      type: object
      required: [sessionId]
      properties:
        sessionId: { type: string }
        imageUrl: { type: string }
        imageBase64: { type: string }
        status: { type: string }
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
...